requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pdfplumber
pydantic
requests
openai